requirements.txt
absl-py
asgiref
astunparse
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
crispy-bootstrap5
cryptography
cycler
defusedxml
distlib
dj-database-url
Django
django-allauth
django-crispy-forms
django-email
django-environ
django-filter
django-model-utils
django-tailwind
djangorestframework
djangorestframework-simplejwt
EasyProcess
entrypoint2
filelock
Flask
flatbuffers
fonttools
fuzzywuzzy
gast
google-auth
google-auth-oauthlib
google-pasta
greenlet
grpcio
h5py
idna
itsdangerous
Jinja2
joblib
keras
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
ml-dtypes
MouseInfo
mss
mysql-connector-python
numpy
oauthlib
opt-einsum
package-name
packaging
pandas
Pillow
pipenv
platformdirs
protobuf
psycopg2
psycopg2-binary
pyasn1
pyasn1-modules
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
PyMySQL
pyparsing
pyperclip
PyRect
pyscreenshot
PyScreeze
python-dateutil
python3-openid
pytweening
pytz
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
soupsieve
SQLAlchemy
sqlparse
tensorboard
tensorboard-data-server
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
typing_extensions
tzdata
urllib3
virtualenv
Werkzeug
wrapt